Partie sans testeur

+++ ft_strtrim.c
@@ -1,3 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strtrim.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aviscogl <aviscogl@student.42lyon.fr>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/11/09 08:35:10 by aviscogl          #+#    #+#             */
+/*   Updated: 2024/11/09 08:35:10 by aviscogl         ###   ########lyon.fr   */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <stdio.h>
+
 int	ft_strlen(const char *s)
 {
 	int	cpt;
@@ -9,12 +23,13 @@ int	ft_strlen(const char *s)
 	}
 	return (cpt);
 }
+/*
 char *ft_strtrim(char const *s1, char const *set)
 {//partir du principe qu il n y a pas de separateur, balayer le set, 
 	char *dest;
-	int len;
-	int cpt;
-	int	sep;
+	size_t len;
+	size_t cpt;
+	size_t	sep;
 
 	sep = 0;
 	len = ft_strlen(s1);
@@ -37,3 +52,4 @@ char *ft_strtrim(char const *s1, char const *set)
 	
 	return (dest);
 }
+*/
